Keep interfaces with their implementations in the same classes
