--- metadata/src/main/java/org/odk/collect/metadata/DeviceDetailsProvider.kt
@@ -1,9 +1,0 @@
-package org.odk.collect.metadata
-
-interface DeviceDetailsProvider {
-    @get:Throws(SecurityException::class)
-    val deviceId: String?
-
-    @get:Throws(SecurityException::class)
-    val line1Number: String?
-}
