--- metadata/src/main/java/org/odk/collect/metadata/InstallIDProvider.kt
@@ -1,5 +1,0 @@
-package org.odk.collect.metadata
-
-interface InstallIDProvider {
-    val installID: String
-}
